refactor(reporter): Move Opossum reporter into its own package
Prepare for splitting up this class.
