+++ reporter/src/funTest/kotlin/reporters/opossum/OpossumReporterFunTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.reporter.reporters
+package org.ossreviewtoolkit.reporter.reporters.opossum
 
 import com.fasterxml.jackson.databind.json.JsonMapper
