+++ reporter/src/main/resources/META-INF/services/org.ossreviewtoolkit.reporter.Reporter
@@ -1,6 +1,5 @@
 org.ossreviewtoolkit.reporter.reporters.CycloneDxReporter
 org.ossreviewtoolkit.reporter.reporters.ExcelReporter
-org.ossreviewtoolkit.reporter.reporters.OpossumReporter
 org.ossreviewtoolkit.reporter.reporters.StaticHtmlReporter
 org.ossreviewtoolkit.reporter.reporters.WebAppReporter
 org.ossreviewtoolkit.reporter.reporters.ctrlx.CtrlXAutomationReporter
@@ -14,4 +13,5 @@ org.ossreviewtoolkit.reporter.reporters.freemarker.asciidoc.ManPageTemplateRepor
 org.ossreviewtoolkit.reporter.reporters.freemarker.asciidoc.PdfTemplateReporter
 org.ossreviewtoolkit.reporter.reporters.freemarker.asciidoc.XHtmlTemplateReporter
 org.ossreviewtoolkit.reporter.reporters.gitlab.GitLabLicenseModelReporter
+org.ossreviewtoolkit.reporter.reporters.opossum.OpossumReporter
 org.ossreviewtoolkit.reporter.reporters.spdx.SpdxDocumentReporter
